+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.35 fix: fix stock_zh_index_hist_csindex interface
+
+    1. 修复 stock_zh_index_hist_csindex 接口
+
 1.6.34 fix: fix fx_spot_quote interface
 
     1. 修复 fx_spot_quote 接口，输出字段
@@ -825,6 +829,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.35 fix: fix stock_zh_index_hist_csindex interface
 
 1.6.34 fix: fix fx_spot_quote interface